% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_algorithms.R
\name{available_algorithms}
\alias{available_algorithms}
\title{List available graph algorithms}
\usage{
available_algorithms(graph)
}
\arguments{
\item{graph}{a `graph` object}
}
\value{
a `tbl_df` of the Graph algorithms available
}
\description{
List the algorithms available for the graph database instance
}
