% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher_code.R
\name{cypher_code}
\alias{cypher_code}
\title{cypher_code.}
\usage{
cypher_code(x)
}
\arguments{
\item{x}{text}
}
\value{
coded text
}
\description{
send secret message.
}
\examples{
xx <- cypher_code(x = "This is a secret coded message")
xx
cypher_decode(xx)
}
