% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decypher.R
\name{decipher}
\alias{decipher}
\title{Decipher}
\usage{
decipher(ciphertext, key)
}
\arguments{
\item{key}{A character vector of length 1 containing the key used to encipher
the ciphertext}

\item{cyihertext}{A character vector of length 1 containing the text to be
deciphered.}
}
\value{
A character vector of length 1 containing the deciphered text
}
\description{
Decipher
}
