% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decypher.R
\name{decypher}
\alias{decypher}
\title{decypher}
\usage{
decypher(cyphertext, key)
}
\arguments{
\item{cyphertext}{A character vector of length 1 containing the text to be
decyphered.}

\item{key}{A character vector of length 1 containing the key used to encypher
the cyphertext}
}
\value{
A character vector of length 1 containing the decyphered text
}
\description{
decypher
}
