% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encypher.R
\name{encipher}
\alias{encipher}
\title{Encipher}
\usage{
encipher(plaintext, key)
}
\arguments{
\item{plaintext}{A character vector of length 1 containing the text to be
enciphered}

\item{key}{A character vector of length 1 containing the key to be used to
encipher the plaintext}
}
\value{
A character vector of length 1 containing the enciphered text
}
\description{
Encipher
}
