% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encypher.R
\name{encypher}
\alias{encypher}
\title{ncypher}
\usage{
encypher(plaintext, key)
}
\arguments{
\item{plaintext}{A character vector of length 1 containing the text to be
encyphered}

\item{key}{A character vector of length 1 containing the key to be used to
encypher the plaintext}
}
\value{
A character vector of length 1 containing the encyphered text
}
\description{
ncypher
}
