% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_node_labels}
\alias{get_node_labels}
\title{get_node_labels}
\usage{
get_node_labels(graph, node_type)
}
\arguments{
\item{graph}{a `graph` object}

\item{node_type}{a character indicating the type of node to query}
}
\value{
a character vector of node labels
}
\description{
get_node_labels
}
