% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_node_types}
\alias{get_node_types}
\title{get_node_types}
\usage{
get_node_types(graph)
}
\arguments{
\item{graph}{a `graph` object}
}
\value{
a character vector of node types
}
\description{
get_node_types
}
