% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_nodes}
\alias{get_nodes}
\title{get_nodes}
\usage{
get_nodes(graph, node_type)
}
\arguments{
\item{graph}{a `graph` object}

\item{node_type}{a character indicating the type of node to return}
}
\value{
a `tibble` of nodes and their properties
}
\description{
get_nodes
}
