% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher.R
\name{get_relationship_labels}
\alias{get_relationship_labels}
\title{Get Relationship Labels}
\usage{
get_relationship_labels(rel)
}
\arguments{
\item{rel}{a character string of the relationship to query}
}
\description{
Get Relationship Labels
}
