% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{get_relationship_types}
\alias{get_relationship_types}
\title{get_relationship_types}
\usage{
get_relationship_types(graph)
}
\arguments{
\item{graph}{a `graph` object}
}
\value{
a character vector of relationship types
}
\description{
get_relationship_types
}
