% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{get_total_nodes}
\alias{get_total_nodes}
\title{get_total_nodes}
\usage{
get_total_nodes(graph)
}
\arguments{
\item{grpah}{a `graph` object}
}
\value{
a numeric value for the total number of nodes in the graph
}
\description{
get_total_nodes
}
