% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{get_total_relationships}
\alias{get_total_relationships}
\title{get_total relatioships}
\usage{
get_total_relationships(graph)
}
\arguments{
\item{grpah}{a `graph` object}
}
\value{
a numeric value for the total number of relationships (edges) in the graph
}
\description{
get_total relatioships
}
