% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_connect.R
\name{graph_connect}
\alias{graph_connect}
\title{Create Graph Object}
\usage{
graph_connect(db, user = NULL, password = NULL, host = NULL)
}
\arguments{
\item{db}{a character string for the database name}

\item{user}{a character string for the database username Default is \code{NULL}. If \code{neo4j_user} environment variable is set, this value willn be used.}

\item{password}{a character string for the user password Default is \code{NULL}. If \code{neo4j_password} environment variable is set, this value willn be used.}

\item{host}{a character string for the database host address. Default is \code{NULL}. If \code{neo4j_host} environment variable is set, this value willn be used.}
}
\value{
a S4 \code{graphdb} object
}
\description{
Create Graph Object
}
