% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{graphdb-class}
\alias{graphdb-class}
\title{graphdb}
\description{
S4 class to store neo4j database connection details
}
\section{Slots}{

\describe{
\item{\code{host}}{a character string for the database host address (IP:PORT)}

\item{\code{user}}{a character string for the database username}

\item{\code{password}}{a character string for the user password}

\item{\code{dbname}}{a character string for the database name}

\item{\code{headers}}{a character string for request headers}
}}

