% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulo_add.R
\name{modulo_add}
\alias{modulo_add}
\title{1-indexed modular addition}
\usage{
modulo_add(i, m, n = 26)
}
\arguments{
\item{i}{Starting number}

\item{m}{Number to be added}

\item{n}{Modulus. Defaults to 26 for alphabetic modular arithmetic}
}
\description{
Provides modular addition with zero excluded.
}
