% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulo_subtract.R
\name{modulo_subtract}
\alias{modulo_subtract}
\title{1-indexed modular subtraction}
\usage{
modulo_subtract(i, m, n = 26)
}
\arguments{
\item{i}{Starting number}

\item{m}{Number to be subtracted}

\item{n}{Modulus. Defaults to 26 for alphabetic modular arithmetic}
}
\description{
Provides modular subtraction with zero excluded.
}
