% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_query.R
\name{parse_query}
\alias{parse_query}
\title{Parse Query Response}
\usage{
parse_query(response)
}
\arguments{
\item{response}{a \code{httr2_response} class}
}
\value{
a \code{tbl_df}
}
\description{
Parse Query Response
}
