% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/run_query-method.R
\name{run_query}
\alias{run_query}
\alias{run_query,graphdb-method}
\title{Run Cypher Query}
\usage{
run_query(graphdb, cypher_query)

\S4method{run_query}{graphdb}(graphdb, cypher_query)
}
\arguments{
\item{graphdb}{a S4 graphdb object}

\item{cypher_query}{a cypher query in the form of a character stringr}
}
\value{
a \code{httr2_response}
}
\description{
Run Cypher Query
}
