% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{delete_nodes}
\alias{delete_nodes}
\title{delete_nodes}
\usage{
delete_nodes(graph, node_type)
}
\arguments{
\item{node_type}{a character indicating the type of node to delete}

\item{grpah}{a `graph` object}
}
\description{
delete_nodes
}
