% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationships.R
\name{get_relationship_labels}
\alias{get_relationship_labels}
\title{get_relationship_labels}
\usage{
get_relationship_labels(graph, rel)
}
\arguments{
\item{graph}{a `graph` object}

\item{rel}{a character indicating the type of relationship to query}
}
\value{
a character vector of relationship labels
}
\description{
get_relationship_labels
}
