% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher.R
\name{get_relationship_labels}
\alias{get_relationship_labels}
\title{#' @export
get_relationship_types <- function()
{
return(paste0('MATCH ()-->() RETURN count(*)'))
}}
\usage{
get_relationship_labels(rel)
}
\description{
#' @export
get_relationship_types <- function()
{
return(paste0('MATCH ()-->() RETURN count(*)'))
}
}
